requirements.txt
PyYAML
pathlib
maigret
ttkthemes